requirements.txt
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
google-auth==2.40.3
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyttsx3==2.99
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
